+++ src/sql_analysis/execute_queries.py
@@ -29,7 +29,6 @@ from src.sql_analysis.plan_analysis.analyse_plans import analyse_plans
 from src.sql_analysis.tools.sql_to_schema import parse_create_table
 from src.sql_analysis.tools.sql_types import base_type_to_duckdb_type, base_type_to_example_value
 
-
 # Define the error table names
 
 
@@ -53,7 +52,6 @@ class IDManager:
         self.n_success = 0
         self.n_error = 0
 
-
     def get_success_id(self) -> int:
         self.success_id += 1
         self.n_success += 1
@@ -63,8 +61,6 @@ class IDManager:
         self.error_id += 1
         self.n_error += 1
         return self.error_id
-
-
 
 
 def make_create_statement(complete_quoted_table_name, columns):
@@ -258,8 +254,16 @@ def populate_tables_with_inserts(repo_id: int, repo_url: str, con: duckdb.DuckDB
     return new_tables
 
 
+def get_table_qualifier(table_name: str, database_schema: Optional[str]) -> str:
+    if database_schema:
+        return f"{quote(database_schema)}.{quote(table_name)}"
+    else:
+        return quote(table_name)
+
 def populate_tables_with_files(repo_id: int, con: duckdb.DuckDBPyConnection,
-                               sandbox_con: duckdb.DuckDBPyConnection, tables: List[Table]):
+                               sandbox_con: duckdb.DuckDBPyConnection, tables: List[Table],
+                               database_schema: str
+                               ):
     data_files = con.execute(f"""
         SELECT table_name, file_url
         FROM {TABLES_DATA_FILES_TABLE_NAME}
@@ -271,11 +275,12 @@ def populate_tables_with_files(repo_id: int, con: duckdb.DuckDBPyConnection,
     for (table_name, file_url) in data_files:
         # check if the table_name exists in the tables
         if quote(table_name.lower()) not in existing_table_names:
-            logging.error(f"Table {table_name} not found in repo {repo_id} for file {file_url} - Existing tables: {existing_table_names}")
+            logging.error(
+                f"Table {table_name} not found in repo {repo_id} for file {file_url} - Existing tables: {existing_table_names}")
             continue
 
         insert_query = f"""
-            INSERT INTO {quote(table_name)}
+            INSERT INTO {get_table_qualifier(table_name, database_schema)}
             SELECT * FROM '{escape_for_insert(file_url)}'
         """
         try:
@@ -285,7 +290,7 @@ def populate_tables_with_files(repo_id: int, con: duckdb.DuckDBPyConnection,
                 f"Failed to insert data from file {file_url} into table {table_name} in repo {repo_id}: {e}")
 
 
-def populate_empty_tables(tables: List[Table], sandbox_con: duckdb.DuckDBPyConnection) -> List[int]:
+def populate_empty_tables(tables: List[Table], sandbox_con: duckdb.DuckDBPyConnection, database_schema: str) -> List[int]:
     """
     Populates empty tables, returning a list of table IDs that were populated.
     """
@@ -294,7 +299,7 @@ def populate_empty_tables(tables: List[Table], sandbox_con: duckdb.DuckDBPyConne
         try:
             logging.info(f"Checking table {table.table_name} with ID {table.table_id} for population")
             # Check if the table is empty
-            count = sandbox_con.execute(f"SELECT COUNT(*) FROM {quote(table.table_name)}").fetchone()[0]
+            count = sandbox_con.execute(f"SELECT COUNT(*) FROM {get_table_qualifier(table.table_name, database_schema)}").fetchone()[0]
             logging.info(f"Table {table.table_name} has {count} rows")
             if count == 0:
                 # insert one valid and one null value into each table to confuse the optimizer
@@ -307,7 +312,7 @@ def populate_empty_tables(tables: List[Table], sandbox_con: duckdb.DuckDBPyConne
                 null_list = ', '.join('NULL' for _ in columns)
 
                 insert_statement = f"""
-                        INSERT INTO {table.table_name} ({columns_list})
+                        INSERT INTO {get_table_qualifier(table.table_name, database_schema)} ({columns_list})
                         VALUES ({values_list}), ({null_list});
                         """
 
@@ -322,7 +327,8 @@ def populate_empty_tables(tables: List[Table], sandbox_con: duckdb.DuckDBPyConne
 
 def execute_query(query_id: int, sql: str, sql_prepared: str, repo_id: int, repo_url: str,
                   con: duckdb.DuckDBPyConnection,
-                  sandbox_con: duckdb.DuckDBPyConnection, tables: List[Table], id_manager: IDManager) -> MockQueryResult:
+                  sandbox_con: duckdb.DuckDBPyConnection, tables: List[Table],
+                  id_manager: IDManager) -> MockQueryResult:
     result: MockQueryResult = try_to_mock_and_execute_query(sandbox_con, sql_prepared, tables)
     if result.was_successful():
         success_id = id_manager.get_success_id()
@@ -342,7 +348,6 @@ def execute_query(query_id: int, sql: str, sql_prepared: str, repo_id: int, repo
             sql = fixed_sql
             sql_prepared = fixed_sql
             return execute_query(query_id, sql, sql_prepared, repo_id, repo_url, con, sandbox_con, tables, id_manager)
-
 
         error_id = id_manager.get_error_id()
         con.execute(f"""
@@ -378,7 +383,6 @@ def execute_queries(repo_id: int, repo_url: str, sandbox_con: duckdb.DuckDBPyCon
 
 
 def get_tables_from_create_statements(repo_id: int, con: duckdb.DuckDBPyConnection):
-
     create_queries = con.execute(f"""
     WITH distrinct_queries AS (
         SELECT sql, id, repo_id, type
@@ -406,7 +410,8 @@ def get_tables_from_create_statements(repo_id: int, con: duckdb.DuckDBPyConnecti
         logging.info(f"No new CREATE TABLE queries found for repo {repo_id}")
         return
 
-    for sql, table_name, query_id, repo_id in tqdm(create_queries, desc="Processing CREATE TABLE queries", unit="query"):
+    for sql, table_name, query_id, repo_id in tqdm(create_queries, desc="Processing CREATE TABLE queries",
+                                                   unit="query"):
         try:
             table_schema = parse_create_table(sql)
             save_table_in_db(con, repo_id, table_schema)
@@ -417,9 +422,9 @@ def get_tables_from_create_statements(repo_id: int, con: duckdb.DuckDBPyConnecti
                 VALUES (?, ?, ?, ?)
             """, (repo_id, query_id, sql, str(e)))
 
-def save_string_column_values(repo_id: int, sandbox_con: duckdb.DuckDBPyConnection, con: duckdb.DuckDBPyConnection,
-                              artificial_populated_ids: List[int]):
 
+def save_string_column_values(repo_id: int, sandbox_con: duckdb.DuckDBPyConnection, con: duckdb.DuckDBPyConnection,
+                              artificial_populated_ids: List[int], database_schema: str):
     print(f"The following tables were artificially populated: {artificial_populated_ids}")
     # get the columns that where recorded in the executable queries and for which we don't have values yet
     columns_to_record = con.execute(f"""
@@ -443,7 +448,7 @@ def save_string_column_values(repo_id: int, sandbox_con: duckdb.DuckDBPyConnecti
         try:
             values_arrow = sandbox_con.execute(f"""
                 SELECT {column_id} as column_id, {quote(column_name)} as value
-                FROM {quote(table_name)} 
+                FROM {get_table_qualifier(table_name, database_schema)}
                 USING SAMPLE {MAX_VALUES_TO_SAVE_PER_COLUMN}
             """).arrow()
 
@@ -464,7 +469,8 @@ def save_string_column_values(repo_id: int, sandbox_con: duckdb.DuckDBPyConnecti
             continue
 
 
-def save_table_counts(repo_id: int, con: duckdb.DuckDBPyConnection, sandbox_con: duckdb.DuckDBPyConnection, artificial_populated_ids: List[int]):
+def save_table_counts(repo_id: int, con: duckdb.DuckDBPyConnection, sandbox_con: duckdb.DuckDBPyConnection,
+                      artificial_populated_ids: List[int], database_schema: str):
     table_counts = con.execute(f"""
         SELECT id, table_name_clean 
         FROM {TABLES_TABLE_NAME} 
@@ -478,7 +484,7 @@ def save_table_counts(repo_id: int, con: duckdb.DuckDBPyConnection, sandbox_con:
             if table_id in artificial_populated_ids:
                 count = 0
             else:
-                count = sandbox_con.execute(f"SELECT COUNT(*) FROM {quote(table_name)}").fetchone()[0]
+                count = sandbox_con.execute(f"SELECT COUNT(*) FROM {get_table_qualifier(table_name, database_schema)}").fetchone()[0]
 
             con.execute(f"""
                 INSERT INTO {TABLE_VALUES_COUNT_TABLE_NAME} (table_id, count)
@@ -530,9 +536,8 @@ def create_views(repo_id: int, repo_url: str, con: duckdb.DuckDBPyConnection,
             f"Processed {n_views} view creation queries in repo {repo_id} ({repo_url}), successfully created {n_success} views, failed to create {n_views - n_success} views.")
 
 
-
-def save_columns_compression_results(repo_id: int, result_path: str, con: duckdb.DuckDBPyConnection, schema: str = '"main"'):
-
+def save_columns_compression_results(repo_id: int, result_path: str, con: duckdb.DuckDBPyConnection,
+                                     schema: str = 'main'):
     # create the compression benchmark results table if it doesn't exist
     con.execute(f"""
         CREATE TABLE IF NOT EXISTS columns_compression_results (
@@ -552,10 +557,10 @@ def save_columns_compression_results(repo_id: int, result_path: str, con: duckdb
 
     # read the csv file
     try:
-        con.execute(f""" 
+        query = f""" 
             WITH columns_full AS (
                 SELECT columns.id as column_id, column_name, table_name, 
-                '{schema}."' || table_name_clean || '"' as full_quoted_table_name,
+                '"{schema}"."' || table_name_clean || '"' as full_quoted_table_name,
                 '"' || column_name || '"' as quoted_column_name
                 FROM {COLUMNS_TABLE_NAME}
                 JOIN {TABLES_TABLE_NAME} ON {COLUMNS_TABLE_NAME}.table_id = {TABLES_TABLE_NAME}.id
@@ -566,8 +571,14 @@ def save_columns_compression_results(repo_id: int, result_path: str, con: duckdb
                 JOIN read_csv_auto('{escape_for_insert(result_path)}') AS results
                     ON lower(results.table) = lower(full_quoted_table_name) 
                     AND lower(results.column) = lower(quoted_column_name)
-            ) INSERT INTO columns_compression_results ( SELECT DISTINCT * FROM joined_data ORDER BY column_id, algorithm )
-        """)
+                ANTI JOIN columns_compression_results ON 
+                    columns_compression_results.column_id = columns_full.column_id 
+                    AND columns_compression_results.algorithm = results.algorithm
+            ) INSERT INTO columns_compression_results ( 
+                SELECT DISTINCT * FROM joined_data ORDER BY column_id, algorithm
+             )
+        """
+        con.execute(query)
     except Exception as e:
         print(f"Failed to read compression benchmark results from {result_path}: {e}")
         return
@@ -579,13 +590,16 @@ def save_columns_compression_results(repo_id: int, result_path: str, con: duckdb
 
 class ExecutionSettings:
 
-    def __init__(self, mode: ExecutionMode, repo_id: Optional[int] = None, query_id: Optional[int] = None, execute: bool = True, collect_statistics: bool = True, compress: bool = True):
+    def __init__(self, mode: ExecutionMode, repo_id: Optional[int] = None, query_id: Optional[int] = None,
+                 execute: bool = True, collect_statistics: bool = True, compress: bool = True,
+                 repo_contains_str: Optional[str] = None):
         self.mode = mode
         self.repo_id = repo_id
         self.query_id = query_id
         self.execute = execute
         self.collect_statistics = collect_statistics
         self.compress = compress
+        self.repo_contains_str = repo_contains_str
 
 
 def execute_repo_queries(settings: ExecutionSettings):
@@ -597,7 +611,7 @@ def execute_repo_queries(settings: ExecutionSettings):
         reset_statistics_tables(con)
 
     con.close()
-    con = get_con(DATABASE_PATH) # reploads some views
+    con = get_con(DATABASE_PATH)  # reploads some views
 
     # if the mode is append we have to get rid of the queries already executed
     append_filter = "AND queries.id NOT IN (SELECT query_id FROM executed_queries_ids)"
@@ -617,6 +631,7 @@ def execute_repo_queries(settings: ExecutionSettings):
             and ({'queries.id = ' + str(query_id) if query_id else 'True'})
             and ({'repos.id = ' + str(repo_id) if repo_id else 'True'})
             and repos.id NOT IN ({', '.join(map(str, EXCLUDED_REPOS))})
+            and {f"repos.repo_name LIKE '%{settings.repo_contains_str}%'" if settings.repo_contains_str else 'True'}
             -- AND '3rd-party-huggingface' IN repos.repo_url  -- only process huggingface for now
             {append_filter if settings.mode == 'append' else ''}
         GROUP BY ALL
@@ -652,50 +667,62 @@ def execute_repo_queries(settings: ExecutionSettings):
 
             create_views(repo_id, repo_url, con, sandbox_con)
 
-            populate_tables_with_files(repo_id, con, sandbox_con, tables)
-            artificial_populated_ids = populate_empty_tables(tables, sandbox_con)
+            # *** SET UP SANDBOX DB WITH DATA ***
+
+            sandbox_con_is_persistent = False
+            if '3rd-party-kaggle' in repo_name:
+                database_schema = repo_name.replace('3rd-party-kaggle-', '')
+                sandbox_con.close()
+                sandbox_database_path = KAGGLE_DATA_DB_PATH
+                sandbox_con = duckdb.connect(sandbox_database_path, read_only=True)  # switch to the kaggle database
+                print(f"Switched to Kaggle database for repo {repo_id} ({repo_name})")
+                sandbox_con_is_persistent = True
+            elif '3rd-party-huggingface' in repo_name:
+                database_schema = repo_name.replace('3rd-party-huggingface-', '')
+                sandbox_con.close()
+                sandbox_database_path = HUGGINFACE_DATA_DB_PATH
+                sandbox_con = duckdb.connect(sandbox_database_path, read_only=True)
+                sandbox_con_is_persistent = True
+            else:
+                database_schema = 'main'
+
+            # *** POPULATE SANDBOX DB WITH DATA IF NEEDED ***
+
+            populate_tables_with_files(repo_id, con, sandbox_con, tables, database_schema)
+            artificial_populated_ids = []
+            if not sandbox_con_is_persistent:
+                artificial_populated_ids = populate_empty_tables(tables, sandbox_con, database_schema)
+
+            # *** EXECUTE QUERIES ***
 
             if settings.execute:
                 execute_queries(repo_id, repo_url, sandbox_con, con, tables, id_manager, query_id)
                 analyse_plans(con, repo_id)
 
-            save_string_column_values(repo_id, sandbox_con, con, artificial_populated_ids)
-            save_table_counts(repo_id, con, sandbox_con, artificial_populated_ids)
+            # *** RECORD STATISTICS ***
 
-            # *** STATISTICS RECORDING ***
-
-            if 'kaggle' in repo_name:
-                database_schema = repo_name.replace('3rd-party-kaggle-', '')
-                sandbox_con.close()
-                sandbox_database_path = KAGGLE_DATA_DB_PATH
-                sandbox_con = duckdb.connect(sandbox_database_path) # switch to the kaggle database
-            if '3rd-party-huggingface' in repo_name:
-                database_schema = clean_name(repo_name.replace('3rd-party-huggingface-', ''))
-                sandbox_con.close()
-                sandbox_database_path = HUGGINFACE_DATA_DB_PATH
-                sandbox_con = duckdb.connect(sandbox_database_path)
-            else:
-                database_schema = ''
+            save_string_column_values(repo_id, sandbox_con, con, artificial_populated_ids, database_schema)
+            save_table_counts(repo_id, con, sandbox_con, artificial_populated_ids, database_schema)
 
             if settings.collect_statistics:
                 record_statistics_for_repo(con, sandbox_con, repo_id, database_schema)
 
             sandbox_con.close()
 
-
             # *** COMPRESSION BENCHMARK ***
 
             if settings.compress:
                 output_path = os.path.join(TMP_DIR, f"compression_benchmark_repo_{repo_id}.csv")
                 try:
-                    run_compression_benchmark(sandbox_database_path, output_path)
-                    save_columns_compression_results(repo_id, output_path, con)
+                    run_compression_benchmark(sandbox_database_path, output_path, database_schema)
+                    save_columns_compression_results(repo_id, output_path, con, database_schema)
                 except Exception as e:
                     print(f"Failed to run compression benchmark for repo {repo_id}: {e}")
 
-            # Close the sandbox connection, save progress, and checkpoint
+            # Delete the sandbox database if needed
             con.execute("CHECKPOINT;")
-            os.remove(sandbox_database_path) # delete the sandbox database file
+            if not sandbox_con_is_persistent:
+                os.remove(sandbox_database_path)  # delete the sandbox database file
 
             # Update counts
             error_count = id_manager.n_error
@@ -709,7 +736,6 @@ def execute_repo_queries(settings: ExecutionSettings):
                 'Success Count': success_count,
                 'Usages': con.execute(f"SELECT COUNT(*) FROM {COLUMN_USAGES_TABLE_NAME}").fetchone()[0],
             })
-
 
     if n_failed_insertions > 0:
         print(
@@ -727,10 +753,10 @@ def execute_repo_queries(settings: ExecutionSettings):
     con.close()
 
 
-
-
 if __name__ == "__main__":
-
     settings = ExecutionSettings(
-        mode='append', execute=False, collect_statistics=False, compress=True, repo_id=41556)
+        mode='append',
+        execute=False, collect_statistics=True, compress=True,
+        repo_id=41735, repo_contains_str=None
+    )
     execute_repo_queries(settings)
